refactor(spec): make Trie fork-agnostic for amsterdam
Move all trie logic to a shared `src/ethereum/trie.py`, removing the
fork-specific `src/ethereum/forks/amsterdam/trie.py`. Also move trie
internal node types (LeafNode, ExtensionNode, BranchNode, InternalNode)
from `ethereum.state` into `ethereum.trie`, and trim `ethereum.state`
to just account types and the PreState protocol.

- New `ethereum/trie.py`: Trie, trie_set/get/copy, root, patricialize,
  encode_node/account/internal_node, EMPTY_TRIE_ROOT, and node types
- `ethereum/state.py`: remove trie node types; import InternalNode from
  ethereum.trie under TYPE_CHECKING to break the runtime cycle
- `amsterdam/trie.py`: deleted; consumers updated to import directly
  from ethereum.trie
- `fork_loader.py`: fall back to ethereum.trie when a fork has no
  fork-specific trie module
